+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+### Changed 
+### Deprecated 
+### Removed
+### Fixed
+### Security
+
+## [1.3.1]
+### Added
 - [issues/50](https://github.com/podaac/l2ss-py/issues/50): Spatial bounds are computed correctly for grouped empty subset operations
 ### Changed 
 - Upgraded `cmr-umm-updater` to 0.2.0